Added data to facilities table
