--- alembic/versions/113614f66c35_creating_first_tables_in_db.py
@@ -1,140 +1,0 @@
-"""Creating first tables in DB
-
-Revision ID: 113614f66c35
-Revises: 
-Create Date: 2024-07-14 22:10:38.367998
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '113614f66c35'
-down_revision: Union[str, None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('facilities',
-                    sa.Column('key', sa.Integer(),
-                              autoincrement=True, nullable=False),
-                    sa.Column('facility_id', sa.String(), nullable=False),
-                    sa.Column('facility_name', sa.String(), nullable=True),
-                    sa.Column('mfl_code', sa.Integer(), nullable=True),
-                    sa.Column('keph_level', sa.String(), nullable=False),
-                    sa.Column('ward', sa.String(), nullable=False),
-                    sa.Column('ward_id', sa.String(), nullable=False),
-                    sa.Column('sub_county', sa.String(), nullable=False),
-                    sa.Column('sub_county_id', sa.String(), nullable=False),
-                    sa.Column('county', sa.String(), nullable=False),
-                    sa.Column('county_id', sa.String(), nullable=True),
-                    sa.PrimaryKeyConstraint('key')
-                    )
-    op.create_index(op.f('ix_facilities_county'),
-                    'facilities', ['county'], unique=False)
-    op.create_index(op.f('ix_facilities_facility_id'),
-                    'facilities', ['facility_id'], unique=True)
-    op.create_index(op.f('ix_facilities_facility_name'),
-                    'facilities', ['facility_name'], unique=False)
-    op.create_index(op.f('ix_facilities_keph_level'),
-                    'facilities', ['keph_level'], unique=False)
-    op.create_index(op.f('ix_facilities_mfl_code'),
-                    'facilities', ['mfl_code'], unique=True)
-    op.create_index(op.f('ix_facilities_sub_county'),
-                    'facilities', ['sub_county'], unique=False)
-    op.create_index(op.f('ix_facilities_ward'),
-                    'facilities', ['ward'], unique=False)
-    op.create_table('products',
-                    sa.Column('product_id', sa.Integer(),
-                              autoincrement=True, nullable=False),
-                    sa.Column('product_name', sa.String(), nullable=False),
-                    sa.Column('product_category', sa.String(), nullable=False),
-                    sa.Column('ven', sa.String(), nullable=False),
-                    sa.Column('is_equipment', sa.Boolean(), nullable=True),
-                    sa.Column('kmfl_classification',
-                              sa.String(), nullable=True),
-                    sa.Column('is_rmnch', sa.Boolean(), nullable=True),
-                    sa.Column('is_in_647_tracer_list',
-                              sa.Boolean(), nullable=True),
-                    sa.PrimaryKeyConstraint('product_id')
-                    )
-    op.create_index(op.f('ix_products_product_category'),
-                    'products', ['product_category'], unique=False)
-    op.create_index(op.f('ix_products_product_id'),
-                    'products', ['product_id'], unique=False)
-    op.create_index(op.f('ix_products_product_name'),
-                    'products', ['product_name'], unique=False)
-    op.create_table('users',
-                    sa.Column('user_id', sa.Integer(),
-                              autoincrement=True, nullable=False),
-                    sa.Column('email', sa.String(), nullable=False),
-                    sa.Column('password', sa.String(), nullable=False),
-                    sa.Column('access_level', sa.String(), nullable=False),
-                    sa.PrimaryKeyConstraint('user_id')
-                    )
-    op.create_index(op.f('ix_users_access_level'), 'users',
-                    ['access_level'], unique=False)
-    op.create_index(op.f('ix_users_email'), 'users', ['email'], unique=True)
-    op.create_index(op.f('ix_users_user_id'), 'users',
-                    ['user_id'], unique=False)
-    op.create_table('allocations',
-                    sa.Column('key', sa.Integer(),
-                              autoincrement=True, nullable=False),
-                    sa.Column('facility_id', sa.String(), nullable=True),
-                    sa.Column('facility_name', sa.String(), nullable=False),
-                    sa.Column('facility_workload',
-                              sa.Integer(), nullable=False),
-                    sa.Column('allocated_amt', sa.Float(), nullable=False),
-                    sa.ForeignKeyConstraint(
-                        ['facility_id'], ['facilities.facility_id'], ),
-                    sa.PrimaryKeyConstraint('key')
-                    )
-    op.create_index(op.f('ix_allocations_facility_name'),
-                    'allocations', ['facility_name'], unique=False)
-    op.create_table('forecasts',
-                    sa.Column('key', sa.Integer(),
-                              autoincrement=True, nullable=False),
-                    sa.Column('product_name', sa.String(), nullable=False),
-                    sa.Column('pack_size', sa.String(), nullable=True),
-                    sa.Column('price_kes', sa.Float(), nullable=True),
-                    sa.Column('quantity_required_for_one_year',
-                              sa.Float(), nullable=True),
-                    sa.Column('facility_id', sa.String(), nullable=True),
-                    sa.Column('product_id', sa.Integer(), nullable=True),
-                    sa.ForeignKeyConstraint(
-                        ['facility_id'], ['facilities.facility_id'], ),
-                    sa.ForeignKeyConstraint(
-                        ['product_id'], ['products.product_id'], ),
-                    sa.PrimaryKeyConstraint('key')
-                    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('forecasts')
-    op.drop_index(op.f('ix_allocations_facility_name'),
-                  table_name='allocations')
-    op.drop_table('allocations')
-    op.drop_index(op.f('ix_users_user_id'), table_name='users')
-    op.drop_index(op.f('ix_users_email'), table_name='users')
-    op.drop_index(op.f('ix_users_access_level'), table_name='users')
-    op.drop_table('users')
-    op.drop_index(op.f('ix_products_product_name'), table_name='products')
-    op.drop_index(op.f('ix_products_product_id'), table_name='products')
-    op.drop_index(op.f('ix_products_product_category'), table_name='products')
-    op.drop_table('products')
-    op.drop_index(op.f('ix_facilities_ward'), table_name='facilities')
-    op.drop_index(op.f('ix_facilities_sub_county'), table_name='facilities')
-    op.drop_index(op.f('ix_facilities_mfl_code'), table_name='facilities')
-    op.drop_index(op.f('ix_facilities_keph_level'), table_name='facilities')
-    op.drop_index(op.f('ix_facilities_facility_name'), table_name='facilities')
-    op.drop_index(op.f('ix_facilities_facility_id'), table_name='facilities')
-    op.drop_index(op.f('ix_facilities_county'), table_name='facilities')
-    op.drop_table('facilities')
-    # ### end Alembic commands ###
