+++ alembic/versions/eb6124c2ad43_adding_more_index.py
@@ -1,0 +1,34 @@
+"""Adding more index
+
+Revision ID: eb6124c2ad43
+Revises: 3c3b1a8f9fd8
+Create Date: 2024-07-15 13:00:38.003557
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = 'eb6124c2ad43'
+down_revision: Union[str, None] = '3c3b1a8f9fd8'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_index(op.f('ix_facilities_county'), 'facilities', ['county'], unique=False)
+    op.create_index(op.f('ix_facilities_keph_level'), 'facilities', ['keph_level'], unique=False)
+    op.create_index(op.f('ix_facilities_sub_county'), 'facilities', ['sub_county'], unique=False)
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_index(op.f('ix_facilities_sub_county'), table_name='facilities')
+    op.drop_index(op.f('ix_facilities_keph_level'), table_name='facilities')
+    op.drop_index(op.f('ix_facilities_county'), table_name='facilities')
+    # ### end Alembic commands ###
