+++ models/facility.py
@@ -8,9 +8,10 @@ class Facility(Base):
     key = Column(Integer, primary_key=True, autoincrement=True)
     facility_id = Column(String, unique=True, index=True, nullable=False)
     facility_name = Column(String, index=True)
-    mfl_code = Column(Integer, unique=True, index=True, nullable=True)
+    facility_ownership = Column(String, nullable=True)
+    mfl_code = Column(Integer, nullable=True)
     keph_level = Column(String, nullable=False, index=True)
-    ward = Column(String, nullable=False, index=True)
+    ward = Column(String, nullable=False)
     ward_id = Column(String, nullable=False)
     sub_county = Column(String, nullable=False, index=True)
     sub_county_id = Column(String, nullable=False)
